check header

+++ print_functions.c
@@ -6,7 +6,7 @@
 /*   By: eoh <eoh@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/02 15:32:33 by eoh               #+#    #+#             */
-/*   Updated: 2022/12/13 13:40:17 by eoh              ###   ########.fr       */
+/*   Updated: 2022/12/13 18:00:45 by eoh              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,11 +67,11 @@ int	print_pointer(va_list ap)
 
 int	print_decimal(va_list ap)
 {
-	int			len;
-	long long	temp;
+	int	len;
+	int	temp;
 
-	temp = va_arg(ap, long long);
-	len = ft_putnbr((int)temp);
+	temp = va_arg(ap, int);
+	len = ft_putnbr(temp);
 	return (len);
 }
